switch trajectory length

+++ RobotArm_MultiServoSweep/InverseKinematics.h
@@ -7,10 +7,13 @@
 #include "Arduino.h"
 //#include "ArmServos.h"
 #include "ServosManager.h"
-#include "RoboArmTurn.h"
+/*
+ #include "RoboArmTurn.h"
+*/
 #include "RampOnce.h"
 //#define DEBUG         //debug logging
 #define BRIEF_LOG     //just a few logs
+
 #define MATH_PI 3.14159265
 
 class InverseKinematics {
@@ -25,12 +28,12 @@ public:
   
   int microsecondsToAngle(double microseconds);
 
-  ArmServoMicrosec moveToAngle_msec(double b, double a1, double a2, int gripSpinAngle, int gripTiltAngle, double g, double duration);
-  ArmServoAngles   moveToAngle     (double b, double a1, double a2, int gripSpinAngle, int gripTiltAngle, double g, double duration);
+  ArmServoMicrosec moveToAngle_msec(double b, double a1, double a2, double gripSpinAngle, double gripTiltAngle, double g, double duration);
+  ArmServoAngles   moveToAngle     (double b, double a1, double a2, double gripSpinAngle, double gripTiltAngle, double g, double duration);
 
   //ArmServoMicrosec moveToAngle2(ArmServoAngles armServoAngles);
 
-  ArmServoAngles moveToPos(double x, double y, double z, int gripSpinAngle, int gripTiltAngle, double g, bool movesScriptEnd, double duration);
+  ArmServoAngles moveToPos(double x, double y, double z, double gripSpinAngle, double gripTiltAngle, double g, bool movesScriptEnd, double duration);
 
   ArmServoAngles moveToPosXYZ(GripPositionXYZ positionXYZ);
