+++ RobotArm_MultiServoSweep/ServosManager.h
@@ -18,12 +18,12 @@ typedef struct {
 }GripPositionXYZ;
 
 typedef struct {
-  int baseAngle;
-  int arm1Angle;
-  int arm2Angle;
-  int gripSpinAngle;
-  int gripTiltAngle;
-  int gripAngle;
+  double baseAngle;
+  double arm1Angle;
+  double arm2Angle;
+  double gripSpinAngle;
+  double gripTiltAngle;
+  double gripAngle;
   double duration;
   bool movesScriptEnd;
 } ArmServoAngles;
